refactor(source): define Parser contract + dedupe raster parser types (stage 2.1)
interface.ts: 新增 Parser<TData,TCfg,TResult> 统一契约 + KnownParsers

13 parser 映射 + KnownParserType 键名联合. 去重 RasterDataType

/ IRGBParseCfg (rgb.ts + ndi.ts 两份相同) 与 IImageCfg

(image.ts) 收敛到 interface.ts 单一来源. factory.ts:

ParserFunction = Parser 替代字面量 (类型擦除版, align cfg? optional).

13 个 parser 文件无需改签名 (隔离 TS 严格模式实测兼容).

对外 API 完全等价, 仅新增类型无运行时改动.

验证: source tsc 31 基线不变, layers tsc 229 不变,

eslint/prettier 通过, source jest 27/27 通过.

BALL_LOG: 闭环 BACKLOG RasterDataType/IRGBParseCfg 重复定义.
