+++ packages/source/src/parser/ndi.ts
@@ -1,27 +1,6 @@
 import type { IParserData } from '@antv/l7-core';
-import type { IRasterCfg } from '../interface';
+import type { IRGBParseCfg, RasterDataType } from '../interface';
 import { extentToCoord } from '../utils/util';
-
-/**
- * @description: 栅格数据解析
- */
-export type RasterDataType =
-  | Uint8Array
-  | Int8Array
-  | Uint16Array
-  | Int16Array
-  | Uint32Array
-  | Int32Array
-  | Float32Array
-  | Float64Array;
-
-export interface IRGBParseCfg extends IRasterCfg {
-  bands?: [number, number, number];
-  countCut?: [number, number];
-  RMinMax?: [number, number];
-  GMinMax?: [number, number];
-  BMinMax?: [number, number];
-}
 
 export default function ndi(data: RasterDataType[], cfg: IRGBParseCfg): IParserData {
   const {
